build: Ignore walk errors.
For example walking a rootfs can give os.Lstat errors on /proc /dev and other
special mount points files.
A better implementation should be to provide a way to the filepath.Walk
function to ignore some patterns. Now the exclusion is done inside the
BuildWalker that is called by Walk after it has already executed the os.Lstat
on the path.